Added LOTS of new logging into logs database, with loads of config options. + Refactored a bit WorldSession::HandleMessagechatOpcode + Made some auction house sql requests into transactions + Fixed some more wrong database loading types + Changed .muteinfo commands output + Removed .npc subname command + Changed "faction" command name to "changefaction" + Misc whatever fixes
Former-commit-id: e3874fef0841c4188abcb133ee6bd08b25cc4fb3

+++ src/game/Channel.cpp
@@ -60,7 +60,7 @@ Channel::Channel(const std::string& name, uint32 channel_id)
         if (result) {
             do {
                 Field *fields = result->Fetch();
-                gmbanned[fields[0].GetUInt64()] = fields[1].GetUInt64();
+                gmbanned[fields[0].GetUInt32()] = fields[1].GetUInt32();
             } while (result->NextRow());
         }
